Update and rename tis_camera.py to imagingsource.py
Hi @kwang428 ,

First, I wanted to apologize for taking so long to get to this pull request. This is great work and we really appreciate your contributions.

I made some small suggested edits to style to conform with PEP and the rest of the package. I changed the name of the file and the class to likewise conform.

I also removed the `width` and `height` variables because the superclass encodes them in the `shape = (height, width)` attribute.

I think the only changes to behavior that I did are:
- remove a `/ 8` in the bitdepth variable. 24 = 3 * 8 implies 8-bit depth.
- fix a comma bug in the else case of the woi function.

Apart from some lingering concerns about the forced-RGB nature of the camera, it looks good.

At this point, I am alright with pushing to main, but if you have time to test the current code on hardware, that would be preferred to pushing slightly untested code.

Let me know what you think!

Best,
Ian