Poetry attempt.
This build is failing to install opencv-python and scipy on my machine due to error messages similar to `Note: This error originates from the build backend, and is likely not a problem with poetry but with scipy (1.13.1) not supporting PEP 517 builds. You can verify this by running 'pip wheel --no-cache-dir --use-pep517 "scipy (==1.13.1)"'.
`

Co-Authored-By: Bodo Kaiser <i@bodokaiser.io>